wrapper content moved back into eventful, tests passing.
